% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_normality.R
\name{check_normality}
\alias{check_normality}
\title{Print skewness, kurtosis, normality test, and draw qqplot and density plot
Note: add doing these by levels}
\usage{
check_normality(x)
}
\arguments{
\item{x}{a vector with of data}
}
\value{

}
\description{
Print skewness, kurtosis, normality test, and draw qqplot and density plot
Note: add doing these by levels
}
