% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_pkgs.R
\name{load_pkgs}
\alias{load_pkgs}
\title{Load all packages I will use for data management.
Note: I think 'import' is the best way to build a package when using external
functions from other packages. I do this only for convenience because most of
the time I use some functions form these packages.}
\usage{
load_pkgs()
}
\value{
none
}
\description{
Load all packages I will use for data management.
Note: I think 'import' is the best way to build a package when using external
functions from other packages. I do this only for convenience because most of
the time I use some functions form these packages.
}
