% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_data_info.R
\name{print_data_info}
\alias{print_data_info}
\title{Print head, summary, and str at the same time}
\usage{
print_data_info(df)
}
\arguments{
\item{df}{a data.frame object}
}
\value{

}
\description{
Print head, summary, and str at the same time
}
