% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Read data using fread}
\usage{
read_data(path_to_data, separator = ",", use_header = TRUE)
}
\arguments{
\item{path_to_data}{The path to the directory contains only data}

\item{separator}{separator of your data file}

\item{use_header}{TRUE if data file contains headers}
}
\value{
a data.frame
}
\description{
Read data using fread
}
