% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_dir.R
\name{setup_dir}
\alias{setup_dir}
\title{Setup the directory}
\usage{
setup_dir(path)
}
\arguments{
\item{path}{The path to your directory}
}
\value{
current directory and the elements inside
}
\description{
Setup the directory
}
